requirements.txt
streamlit
pandas
pytesseract
pdf2image
pillow
openpyxl
